pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "speechify-client"
version = "0.1.0"
description = "A simple, type-safe REST client for the Speechify AI API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "rich>=13.0.0",
]

[tool.ruff]
line-length = 119
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["speechify_client"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
